+++ src/main/java/com/example/ecom/dto/user/UserRequest.java
@@ -1,4 +1,4 @@
-package com.example.ecom.dto.userDTO;
+package com.example.ecom.dto.user;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
@@ -42,17 +42,17 @@ public class UserRequest {
     @NotBlank(message = "Address is required!")
     @NotNull(message = "Address is required!")
     private String address;
-    
+
     @NotEmpty(message = "First name is required!")
     @NotBlank(message = "First name is required!")
     @NotNull(message = "First name is required!")
     private String firstName;
-    
+
     @NotEmpty(message = "Last name is required!")
     @NotBlank(message = "Last name is required!")
     @NotNull(message = "Last name is required!")
     private String lastName;
-    
+
     @NotNull(message = "Email is required!")
     @Pattern(regexp = TypeValidation.EMAIL, message = "Email is invalid!")
     private String email;
