+++ src/main/java/com/example/ecom/repository/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.example.ecom.repository.userRepository;
+package com.example.ecom.repository.user;
 
 import java.util.List;
 import java.util.Map;
@@ -6,8 +6,9 @@ import java.util.Optional;
 
 public interface UserRepository {
     Optional<List<User>> getUsers(Map<String, String> allParams, String keySort, int page, int pageSize,
-                        String sortField);
+            String sortField);
 
     void insertAndUpdate(User user);
+
     long getTotalPage(Map<String, String> allParams);
 }
